[CF] AUDIT-30 Создать сервис для работы с типами событий (audit_event_type). исправление после CR https://jira.i-novus.ru/browse/AUDIT-30

+++ audit-api/src/main/java/ru/i_novus/ms/audit/model/EventTypeCriteria.java
@@ -21,5 +21,5 @@ public class EventTypeCriteria extends RestCriteria {
 
     @ApiParam("Идентификатор типа журнала")
     @QueryParam("auditTypeId")
-    private Integer auditTypeId;
+    private Short auditTypeId;
 }
